Move all inline components to separate files
